optimized data extraction process
